# src/main/java/com/zestt/assign/DTO/Response/AuthResponse.java
package com.zestt.assign.DTO.Response;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class AuthResponse {
 private String accessToken;
    private String refreshToken;
}