# src/main/java/com/zestt/assign/DTO/Response/ItemResponse.java
package com.zestt.assign.DTO.Response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ItemResponse {
    private Integer id;
    private Integer quantity;
}