# src/main/java/com/zestt/assign/DTO/Response/ProductResponse.java
package com.zestt.assign.DTO.Response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ProductResponse {

    private Integer id;
    private String productName;
}