# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Zest Assignment API
  version: 1.0.0
  description: API documentation generated from current Spring Boot controllers.
servers:
  - url: http://localhost:8080
tags:
  - name: Auth
  - name: Products
  - name: Items
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            text/plain:
              schema:
                type: string
                example: User registered successfully
        '404':
          description: User-related error
          content:
            text/plain:
              schema:
                type: string
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Auth success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /api/v1/products/create:
    post:
      tags: [Products]
      summary: Create product
      security:
        - bearerAuth: []
      description: Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/products/{id}:
    get:
      tags: [Products]
      summary: Get product by id
      security:
        - bearerAuth: []
      description: Requires USER or ADMIN role.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Products]
      summary: Update product
      security:
        - bearerAuth: []
      description: Requires ADMIN role.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Products]
      summary: Delete product
      security:
        - bearerAuth: []
      description: Requires ADMIN role.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/products/{productId}/items:
    get:
      tags: [Items]
      summary: Get items by product
      security:
        - bearerAuth: []
      description: Requires USER or ADMIN role.
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Items]
      summary: Create item under product
      security:
        - bearerAuth: []
      description: Requires ADMIN role.
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [userName, password]
      properties:
        userName:
          type: string
        password:
          type: string
    RegisterRequest:
      type: object
      required: [userName, password]
      properties:
        userName:
          type: string
        password:
          type: string
        role:
          type: string
          description: Accepts USER or ADMIN (mapped internally to ROLE_USER/ROLE_ADMIN).
          enum: [USER, ADMIN]
    ProductRequest:
      type: object
      required: [productName]
      properties:
        productName:
          type: string
          maxLength: 255
    ItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 0
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
        productName:
          type: string
    ItemResponse:
      type: object
      properties:
        id:
          type: integer
        quantity:
          type: integer
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string